PythonClient/python_files/requirements.txt
absl-py==0.7.0
astor==0.7.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
gast==0.2.2
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
grpcio==1.16.1
h5py==2.9.0
httplib2==0.12.1
idna==2.8
Jinja2==2.9.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.8
matplotlib==3.0.2
mkl-fft==1.0.10
mkl-random==1.0.2
numpy==1.15.4
oauth2client==4.1.3
olefile==0.44
pandas==0.24.1
pick==0.6.4
Pillow==4.0.0
prettytable==0.7.2
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pyfiglet==0.8.post1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.0
Werkzeug==0.14.1
